my-chalkboard-cash/Backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
httpx==0.27.2
Flask
gunicorn
pandas
numpy
opencv-python
scikit-learn
yfinance
